Corrected a bug in xdg-get-user-file.
A bug was present in xdg-get-user-file, that duplicated the emacs
subdirectory twice when being called by the locate-user-emacs-*
functions.  The bug arose from the passage of the xdg-*-dir directories
to the user-emacs-*-dir, which already contained the subdirectory.  The
function was then duplicating it.  Now the function does not insert
'emacs' as a subdirectory anymore.